[feat/jmt_search_edit_text]: View 관련 extension 분리 및 fadeIn/Out 애니메이션 추가
